🚚 Moved Go to LI Profile CTA + added callback prop
